+++ frontend/app/apis/[hashId]/page.tsx
@@ -4,7 +4,6 @@ import Document from "./Document";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import callApi from "@/utils/callApi";
 import { ApiGetResponse } from "gpinterface-shared/type/api";
-import { cn } from "@/utils/css";
 import useUserStore from "@/store/user";
 import { MessageSquareCode, SquareCode } from "lucide-react";
 import IconTextButton from "@/components/buttons/IconTextButton";
@@ -30,8 +29,12 @@ export default function Page({ params }: { params: { hashId: string } }) {
   }, [hashId]);
 
   const [tab, setTab] = useState("chat");
-  const isHidden = useCallback(
-    (_tab: string) => (tab === _tab ? "" : "hidden"),
+  const getTabContentClassName = useCallback(
+    (_tab: string) => {
+      const className = "w-full h-full pt-9 md:pt-0 overflow-hidden";
+      if (tab === _tab) return className;
+      return className + " hidden";
+    },
     [tab]
   );
 
@@ -45,7 +48,7 @@ export default function Page({ params }: { params: { hashId: string } }) {
     <div className="w-full flex-1 flex flex-col gap-3 pt-3 overflow-hidden">
       <div className="px-3 whitespace-pre-wrap">{api?.description ?? ""}</div>
       <div className="flex-1 grid grid-cols-[1fr_auto] overflow-hidden relative">
-        <div className="absolute top-0 left-3">
+        <div className="absolute top-0 left-3 flex md:flex-col gap-3">
           <IconTextButton
             onClick={() => setTab("chat")}
             className="w-24 md:w-32"
@@ -54,8 +57,6 @@ export default function Page({ params }: { params: { hashId: string } }) {
             selected={tab === "chat"}
             responsive
           />
-        </div>
-        <div className="absolute top-0 md:top-[2.75rem] left-[7.5rem] md:left-3">
           <IconTextButton
             onClick={() => setTab("document")}
             className="w-28 md:w-32"
@@ -70,12 +71,7 @@ export default function Page({ params }: { params: { hashId: string } }) {
           useApi={[api, setApi]}
           editable={editable}
         />
-        <div
-          className={cn(
-            "w-full h-full pt-9 md:pt-0 overflow-hidden",
-            isHidden("chat")
-          )}
-        >
+        <div className={getTabContentClassName("chat")}>
           {!!api && (
             <Contents
               chat={api.chat}
@@ -85,12 +81,7 @@ export default function Page({ params }: { params: { hashId: string } }) {
             />
           )}
         </div>
-        <div
-          className={cn(
-            "w-full flex-1 pt-9 md:pt-0 overflow-hidden",
-            isHidden("document")
-          )}
-        >
+        <div className={getTabContentClassName("document")}>
           <Document api={api} />
         </div>
         <ModelPanel topPadding={false}>
